requirements.txt
pillow
PyPDF2
fpdf